Change Blox usage types. ab#227

+++ src/Client/Shared/Models/Blox.cs
@@ -6,12 +6,12 @@
         public string? Name { get; set; }
         public string? OwnerId { get; set; }
         public string? PoolName { get; set; }
-        public float? AvailableSpace { get; set; }
-        public float? CurrentUse { get; set; }
-        public float? PhotosUsed { get; set; }
-        public float? VideosUsed { get; set; }
-        public float? AudiosUsed { get; set; }
-        public float? DocsUsed { get; set; }
-        public float? OtherUsed { get; set; }
+        public long? AvailableSpace { get; set; }
+        public long? CurrentUse { get; set; }
+        public long? PhotosUsed { get; set; }
+        public long? VideosUsed { get; set; }
+        public long? AudiosUsed { get; set; }
+        public long? DocsUsed { get; set; }
+        public long? OtherUsed { get; set; }
     }
 }
